+++ src/WebDriver.php
@@ -90,10 +90,10 @@ class WebDriver extends HttpDriver
     /**
      * Send a typing indicator and wait for the given amount of seconds.
      * @param IncomingMessage $matchingMessage
-     * @param int $seconds
+     * @param float $seconds
      * @return mixed
      */
-    public function typesAndWaits(IncomingMessage $matchingMessage, int $seconds)
+    public function typesAndWaits(IncomingMessage $matchingMessage, float $seconds)
     {
         $this->replies[] = [
             'message' => TypingIndicator::create($seconds),
